+++ packages/json/src/index.ts
@@ -1,2 +1,2 @@
-export * from './Error/JSONAError';
 export * from './JSONA';
+export * from './JSONAError';
