+++ packages/random/src/RandomError.ts
