+++ packages/random/src/index.ts
@@ -1,1 +1,2 @@
 export * from './Random';
+export * from './RandomError';
